\name{genesets}
\alias{genesets}
\docType{data}
\title{Names of genesets used in pathprint
}
\description{
An index to the genesets used in pathprint for each species, referenced by 
common and latin name
}
\usage{genesets}

\references{
Altschuler, G. M., O. Hofmann, I. Kalatskaya, R. Payne, S. J. Ho Sui, U. Saxena,
A. V. Krivtsov, S. A. Armstrong, T. Cai, L. Stein and W. A. Hide (2013). 
"Pathprinting: An integrative approach to understand the functional basis of 
disease." Genome Med 5(7): 68.
}
\examples{
data("genesets")
genesets
}
\keyword{datasets}
