\name{genesets}
\alias{genesets}
\docType{data}
\title{Names of genesets used in pathprint
}
\description{
An index to the genesets used in pathprint for each species, referenced by common and latin name
}
\usage{genesets}

\references{
Pathway activation mapping for large scale comparison of gene expression 
Altschuler G, Hofmann O, Payne R, Kalatskaya I, Stein L, Cai T, Hide W \emph{in preparation}
}
\examples{
genesets
}
\keyword{datasets}
